+++ project/src/components/base/ButtonDelete.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="btn-group">
+    <button class="delete btn btn-danger" data-bs-toggle="modal" data-bs-target="#unregisterModal">
+      회원 탈퇴
+    </button>
+  </div>
+</template>
+
+<script setup></script>
+
+<style scoped>
+.delete .btn {
+  width: 7rem;
+  border: none;
+  border-radius: 2px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.05);
+}
+</style>
